% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vam_ado.R
\name{vam}
\alias{vam}
\title{Estimate Teacher Value-Added}
\usage{
vam(
  y = NULL,
  teacher = NULL,
  year = NULL,
  class = NULL,
  controls = NULL,
  by = NULL,
  tfx_resid = NULL,
  absorb = NULL,
  driftlimit = NULL,
  quasi = FALSE,
  data = NULL,
  return_df_only = FALSE,
  tv_name = "tv",
  scores_name = "score_r"
)
}
\arguments{
\item{y}{Required. The outcome variable in quotations.}

\item{teacher}{Required. The name of the teacher identifier in quotations.}

\item{year}{Required. The name of the year identifier in quotations.}

\item{class}{Required. The name of the classroom identifier in quotations.}

\item{controls}{Controls to be used in the residualization.}

\item{by}{Perform VA estimation separately for each by-group. Ex by = c("subject", "level")}

\item{tfx_resid}{Absorb fixed effects during y residualization, but include those fixed effects in the residual}

\item{absorb}{Residualize y on absorbed fixed effects}

\item{driftlimit}{estimate only # autocovariances; set all further autocovariances equal to the last estimate}

\item{quasi}{Currently not implemented.}

\item{data}{Required. Data frame to use.}

\item{return_df_only}{Default FALSE. If TRUE, return only the original dataframe with value-added and residuals appended. If FALSE,
return a list of (a) the original df with value-added and residuals, (b) the variance estimates, (c) the autocovariances by lag, and
(d) a teacher-year-(by) dataset with the value-added estimates}

\item{tv_name}{Default "tv". String denoting the name of the value-added forecast}

\item{scores_name}{Default "score_r". String denoting the name of the score residual.}
}
\value{
A dataframe or list depending on \code{return_df_only}.
}
\description{
This program uses student-level outcomes (typically test scores) to compute teacher value-added (VA) measures.
The estimation procedure accounts for drift in teacher VA over time.
This package is based on Michael Stepner's Stata program and intended to produce identical results with the speed gains made possible by R.
}
\examples{
\donttest{
returned <- vam(
  by = c("lvl", "subject"), 
  data = df_prepped, 
  controls = controls, 
  teacher = "mepid",
  class = "section_id",
  year = "syear",
  tfx_resid = "mepid", 
  driftlimit = 7,
  y = "test",
  return_df_only = FALSE,
  tv_name = "tv",
  scores_name = "score_r"
)
}
}
